+++ sistema_advocacia/src/main/java/com/sistema_advocacia/repository/ServicoJuridicoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServicoJuridicoRepository extends JpaRepository<ServicoJuridico, Long> {
+public interface ServicoJuridicoRepository extends JpaRepository<ServicoJuridico, Integer> {
 }
